+++ package.json
@@ -11,26 +11,40 @@
   "files": [
     "dist"
   ],
+  "exports": {
+    ".": "./dist/index.js",
+    "./define": "./dist/index.js",
+    "./tab-container": "./dist/tab-container-element.js",
+    "./tab-container/define": "./dist/tab-container-element-define.js"
+  },
   "scripts": {
     "clean": "rm -rf dist",
-    "lint": "eslint src/*.ts test/*.js",
+    "lint": "eslint . --ext .js,.ts && tsc --noEmit",
+    "lint:fix": "eslint --fix . --ext .js,.ts",
     "prebuild": "npm run clean && npm run lint && mkdir dist",
-    "build": "tsc",
+    "bundle": "esbuild --bundle dist/index.js --keep-names --outfile=dist/bundle.js --format=esm",
+    "build": "tsc && npm run bundle && npm run manifest",
     "pretest": "npm run build",
-    "test": "web-test-runner test/test.js",
+    "test": "web-test-runner",
     "prepublishOnly": "npm run build",
-    "postpublish": "npm publish --ignore-scripts --@github:registry='https://npm.pkg.github.com'"
+    "postpublish": "npm publish --ignore-scripts --@github:registry='https://npm.pkg.github.com'",
+    "manifest": "custom-elements-manifest analyze"
   },
   "prettier": "@github/prettier-config",
   "devDependencies": {
-    "@github/prettier-config": "0.0.4",
-    "@open-wc/testing": "^3.1.6",
-    "@web/dev-server-esbuild": "^0.3.2",
-    "@web/test-runner": "^0.14.0",
-    "eslint": "^8.23.1",
-    "eslint-plugin-github": "^4.3.7",
-    "typescript": "^4.4.3"
+    "@custom-elements-manifest/analyzer": "^0.8.3",
+    "@github/prettier-config": "^0.0.6",
+    "@open-wc/testing": "^3.2.0",
+    "@web/dev-server-esbuild": "^0.4.1",
+    "@web/test-runner": "^0.16.1",
+    "@web/test-runner-playwright": "^0.10.1",
+    "esbuild": "^0.18.3",
+    "eslint": "^8.42.0",
+    "eslint-plugin-custom-elements": "^0.0.8",
+    "eslint-plugin-github": "^4.8.0",
+    "typescript": "^5.1.3"
   },
+  "customElements": "custom-elements.json",
   "eslintIgnore": [
     "dist/"
   ]
