+++ src/tab-container-element-define.ts
@@ -1,0 +1,33 @@
+import {TabContainerElement} from './tab-container-element.js'
+
+const root = (typeof globalThis !== 'undefined' ? globalThis : window) as typeof window
+try {
+  root.TabContainerElement = TabContainerElement.define()
+} catch (e: unknown) {
+  if (
+    !(root.DOMException && e instanceof DOMException && e.name === 'NotSupportedError') &&
+    !(e instanceof ReferenceError)
+  ) {
+    throw e
+  }
+}
+
+type JSXBase = JSX.IntrinsicElements extends {span: unknown}
+  ? JSX.IntrinsicElements
+  : Record<string, Record<string, unknown>>
+declare global {
+  interface Window {
+    TabContainerElement: typeof TabContainerElement
+  }
+  interface HTMLElementTagNameMap {
+    'tab-container': TabContainerElement
+  }
+  namespace JSX {
+    interface IntrinsicElements {
+      ['tab-container']: JSXBase['span'] & Partial<Omit<TabContainerElement, keyof HTMLElement>>
+    }
+  }
+}
+
+export default TabContainerElement
+export * from './tab-container-element.js'
