+++ tsconfig.json
@@ -1,17 +1,13 @@
 {
-  "include": ["src"],
   "compilerOptions": {
+    "module": "es2020",
+    "target": "es2017",
     "strict": true,
-    "module": "ESNext",
-    "esModuleInterop": true,
-    "lib": ["es6", "dom", "dom.iterable"],
-    "target": "ES2020",
     "moduleResolution": "node",
     "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "forceConsistentCasingInFileNames": true,
-    "rootDir": "src",
-    "outDir": "dist"
-  }
+    "outDir": "dist",
+    "lib": ["dom", "dom.iterable", "es2020"],
+    "removeComments": true
+  },
+  "files": ["src/index.ts"]
 }
